+++ src/main/webapp/WEB-INF/jsp/00-js-snippets/datepicker-future.js
@@ -1,0 +1,1 @@
+$(".datepicker").datepicker({ firstDay:1, dateFormat:'yy-mm-dd', minDate:0 , maxDate:null});