+++ src/main/webapp/WEB-INF/jsp/00-js-snippets/datepicker.js
@@ -1,1 +1,0 @@
-$(".datepicker").datepicker({ firstDay:1, dateFormat:'yy-mm-dd', minDate:0 });